1-string_nconcat

+++ 0x0C-more_malloc_free/1-string_nconcat.c
@@ -12,8 +12,9 @@
 char *string_nconcat(char *s1, char *s2, unsigned int n)
 {
 	char *str;
-	unsigned int len1 = 0;
-	unsigned int len2 = 0;
+
+	size_t len1 = 0;
+	size_t len2 = 0;
 	unsigned int i = 0;
 
 	if (s1 == NULL)
@@ -32,6 +33,7 @@ char *string_nconcat(char *s1, char *s2, unsigned int n)
 		n = len2;
 
 	str = malloc(sizeof(len1 + n + 1));
+
 	if (str == NULL)
 		return (NULL);
 
@@ -42,7 +44,5 @@ char *string_nconcat(char *s1, char *s2, unsigned int n)
 		str[len1 + i] = s2[i];
 
 	str[len1 + n] = '\0';
-
 	return (str);
 }
-
